THE-438 split S3 compat e2e tests
